+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Add pygame
+- Add new requirement - pygame
+- Create a loading screen and a game field
+- Create Tile and TileManager classes to process the game field
+- Modify the main loop to include the game start and end
+
 ## Add hand recognition
 - Create CHANGELOG.md
 - Add two new requirements - mediapipe and numpy